+++ src/lipsumpy/lipsum_cpp/exceptions.py
@@ -1,0 +1,5 @@
+"""Exceptions defined for lipsum-cpp."""
+
+class LipsumCppError(Exception):
+    """An exception originating from the C++ library."""
+    pass